requirements.txt
streamlit
groq
python-dotenv
requests
pytest
pytest-html